+++ phase1/src/model/products/NoteBook.java
@@ -4,7 +4,7 @@ public class NoteBook extends Stationery {
     private int numOfSheets;
     private String paperType;
 
-    public NoteBook( String productName, int productPrice, int numOfProduct, String manufactureCountry, int numOfSheets, String paperType) {
+    public NoteBook( String productName, double productPrice, int numOfProduct, String manufactureCountry, int numOfSheets, String paperType) {
         super( productName, productPrice, numOfProduct, manufactureCountry);
         this.numOfSheets = numOfSheets;
         this.paperType = paperType;
