+++ phase1/src/model/products/Pen.java
@@ -3,7 +3,7 @@ package model.products;
 public class Pen extends Stationery {
     private PenColor penColor;
 
-    public Pen( String productName, int productPrice, int numOfProduct, String manufactureCountry, String penColor) {
+    public Pen( String productName, double productPrice, int numOfProduct, String manufactureCountry, String penColor) {
         super( productName, productPrice, numOfProduct, manufactureCountry);
         this.penColor = PenColor.valueOf(penColor);
     }
